// src/core/interfaces/random.ts
/**
 * This file was automatically generated by joi-to-typescript
 * Do not modify this file manually
 */

export interface Random {
  length?: number;
  max?: number;
  min?: number;
  precision?: number;
  start?: number;
  type: 'string' | 'integer' | 'boolean' | 'float' | 'hexaDecimal' | 'word' | 'words' | 'image';
}